requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.5.2
asttokens==2.4.1
async-timeout==5.0.1
attrs==24.2.0
backcall==0.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.1.1
cycler==0.12.1
datasets==2.14.4
debugpy==1.8.8
decorator==5.1.1
dill==0.3.7
efficientnet-pytorch==0.7.1
exceptiongroup==1.2.2
executing==2.1.0
facenet-pytorch==2.5.3
fastapi==0.115.5
filelock==3.16.1
fonttools==4.55.0
frozenlist==1.5.0
fsspec==2024.10.0
h11==0.14.0
huggingface-hub==0.26.2
idna==3.10
importlib-metadata==8.5.0
importlib-resources==6.4.5
ipykernel==6.29.5
ipython==8.12.3
jedi==0.19.2
jinja2==3.1.4
joblib==1.4.2
jupyter-client==8.6.3
jupyter-core==5.7.2
kiwisolver==1.4.7
line-profiler==4.1.3
markupsafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.15
nest-asyncio==1.6.0
networkx==3.1
numpy==1.24.4
packaging==24.2
pandas==2.0.3
parso==0.8.4
pickleshare==0.7.5
pillow==10.4.0
pip==24.3.1
platformdirs==4.3.6
prompt-toolkit==3.0.48
propcache==0.2.0
psutil==6.1.0
pure-eval==0.2.3
pyarrow==17.0.0
pydantic==2.10.1
pydantic-core==2.27.1
pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
pywin32==308
pyyaml==6.0.2
pyzmq==26.2.0
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.3.2
scipy==1.10.1
setuptools==75.3.0
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.41.3
sympy==1.13.3
threadpoolctl==3.5.0
tokenizers==0.20.3
tornado==6.4.1
tqdm==4.67.0
traitlets==5.14.3
transformers==4.46.3
typing-extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.1
wcwidth==0.2.13
wheel==0.45.0
xxhash==3.5.0
yarl==1.15.2
zipp==3.20.2
